feat : added getTimestampToEvent api and reorganised tests to limits siez of index.test.ts
